donts.store/requirements.txt
aldryn-apphooks-config==0.4.1
aldryn-boilerplates==0.7.7
aldryn-categories==1.1.0
aldryn-common==1.0.4
aldryn-forms==3.0.4
aldryn-newsblog==2.1.1
aldryn-people==2.1.0
aldryn-translation-tools==0.2.1
Babel==2.6.0
backcall==0.1.0
backport-collections==0.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
cmsplugin-contact==1.1.3
cmsplugin-contact-plus==1.3.12
cssselect==1.0.3
cssutils==1.0.2
decorator==4.3.0
defusedxml==0.5.0
Django==1.11.15
django-absolute==0.3
django-admin-sortable==2.1.9
django-admin-sortable2==0.6.19
django-allauth==0.35.0
django-angular==2.0.3
django-appconf==1.0.2
django-appdata==0.2.0
django-classy-tags==0.8.0
django-cms==3.4.6
django-compressor==2.2
django-connected==0.1.5
django-contact-form==1.6
django-emailit==0.2.3
django-envelope==1.3
django-filer==1.3.2
django-filter==1.1.0
django-formtools==2.1
django-fsm==2.6.0
django-fsm-admin==1.2.4
django-haystack==2.5.0
django-ipware==1.1.1
django-mptt==0.8.7
django-object-tools==1.11.0
django-parler==1.9.2
django-phonenumber-field==1.3.0
django-polymorphic==1.3
django-post-office==3.0.3
django-ranged-response==0.2.0
django-recaptcha==1.4.0
django-redis-cache==1.7.1
django-redis-sessions==0.6.1
django-rest-auth==0.9.2
django-sass-processor==0.6
django-sekizai==0.10.0
django-select2==6.0.3
-e git+https://github.com/awesto/django-shop@08c88ee12da3b3d04aa70af82518116b893e5333#egg=django_shop
django-simple-captcha==0.5.9
django-simple-math-captcha==1.0.7
django-sizefield==0.9.1
django-sortedm2m==1.5.0
django-tablib==3.2
django-taggit==0.23.0
django-treebeard==4.2.0
djangocms-admin-style==1.2.8
djangocms-attributes-field==0.3.0
djangocms-bootstrap3==0.3.2
djangocms-cascade==0.16
djangocms-forms==0.2.5
djangocms-googlemap==1.1.1
djangocms-instagram==0.1.1
djangocms-text-ckeditor==3.4.0
djangorestframework==3.7.7
djangoshop-stripe==0.12.3
drf-haystack==1.6.1
easy-thumbnails==2.5
elasticsearch==1.7.0
enum34==1.1.6
et-xmlfile==1.0.1
hashids==1.2.0
hiredis==0.2.0
html5lib==0.9999999
httplib2==0.11.3
idna==2.6
ipython==5.8.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.12.1
jsonfield==1.0.3
libsass==0.12.3
lxml==4.2.4
oauthlib==1.0.3
odfpy==1.3.6
olefile==0.45.1
openpyxl==2.4.9
parso==0.3.1
pathlib2==2.3.2
pexpect==4.6.0
phonenumbers==8.9.12
phonenumberslite==8.9.12
pickleshare==0.7.4
Pillow==4.3.0
pkg-resources==0.0.0
premailer==3.2.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
python-dateutil==2.6.0
python-instagram==1.3.2
python-openid==2.2.5
python-slugify==1.2.0
python3-openid==3.1.0
pytz==2015.7
PyYAML==3.13
rcssmin==1.0.6
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.5.0
rjsmin==1.0.12
scandir==1.9.0
simplegeneric==0.8.1
simplejson==3.16.0
six==1.11.0
South==1.0.2
stripe==1.53.0
tablib==0.12.1
traitlets==4.3.2
unicodecsv==0.14.1
Unidecode==0.4.18
urllib3==1.22
virtualenv==13.1.2
wcwidth==0.1.7
xlrd==1.1.0
xlwt==1.3.0
YURL==0.13
